+++ framework/Furion/SpecificationDocument/Options/SpecificationDocumentSettingsOptions.cs
@@ -96,7 +96,7 @@ namespace Furion.SpecificationDocument
             options.DocExpansionState ??= DocExpansion.List;
 
             // 加载项目注册和模块化/插件注释
-            var projectXmlComments = App.Assemblies.Where(u => u.GetName().Name != "Furion").Select(t => t.GetName().Name);
+            var projectXmlComments = App.Assemblies.Where(u => u.GetName().Name != nameof(Furion)).Select(t => t.GetName().Name);
             var externalXmlComments = App.ExternalAssemblies.Any() ? App.Settings.ExternalAssemblies.Select(u => u.EndsWith(".dll") ? u[0..^4] : u) : Array.Empty<string>();
             XmlComments ??= projectXmlComments.Concat(externalXmlComments).ToArray();